+++ tests/expressions_sparse.cpp
@@ -2184,7 +2184,7 @@ BOOST_AUTO_TEST_CASE(outer_product) {
 BOOST_AUTO_TEST_CASE(dot) {
   // Test the dot expression function
   int result = 0;
-  BOOST_REQUIRE_NO_THROW(result = a("a,b,c") * b("a,b,c"));
+  BOOST_REQUIRE_NO_THROW(result = static_cast<int>(a("a,b,c") * b("a,b,c")));
   BOOST_REQUIRE_NO_THROW(result += a("a,b,c") * b("a,b,c"));
   BOOST_REQUIRE_NO_THROW(result -= a("a,b,c") * b("a,b,c"));
   BOOST_REQUIRE_NO_THROW(result *= a("a,b,c") * b("a,b,c"));
@@ -2210,7 +2210,7 @@ BOOST_AUTO_TEST_CASE(dot_permute) {
   Permutation perm({2, 1, 0});
   // Test the dot expression function
   int result = 0;
-  BOOST_REQUIRE_NO_THROW(result = a("a,b,c") * b("c,b,a"));
+  BOOST_REQUIRE_NO_THROW(result = static_cast<int>(a("a,b,c") * b("c,b,a")));
   BOOST_REQUIRE_NO_THROW(result += a("a,b,c") * b("c,b,a"));
   BOOST_REQUIRE_NO_THROW(result -= a("a,b,c") * b("c,b,a"));
   BOOST_REQUIRE_NO_THROW(result *= a("a,b,c") * b("c,b,a"));
@@ -2236,7 +2236,7 @@ BOOST_AUTO_TEST_CASE(dot_permute) {
 BOOST_AUTO_TEST_CASE(dot_expr) {
   // Test the dot expression function
   int result = 0;
-  BOOST_REQUIRE_NO_THROW(result = a("a,b,c") * b("a,b,c"));
+  BOOST_REQUIRE_NO_THROW(result = static_cast<int>(a("a,b,c") * b("a,b,c")));
 
   // Compute the expected value for the dot function.
   int expected = 0;
